requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126
albumentations==2.0.6
matplotlib==3.10.1
numpy==2.2.5
Pillow==11.2.1
pycocotools==2.0.8
rich==14.0.0
tifffile==2025.3.30
torch==2.7.0+cu126
torchvision==0.22.0+cu126
